pyproject.toml
[project]
name = "palatepilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "beautifulsoup4>=4.13.4",
    "googlesearch-python>=1.3.0",
    "julep>=2.17.0",
    "markdown>=3.8.2",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
]
